refactor(post): split service.go into per-domain service files
Break the 732-line service.go into service_post.go, service_like.go,
service_category.go, service_tag.go and service_moderation.go, all flat
inside the internal/post package. Each file keeps only the imports it
uses and its related query/request types next to the functions that use
them.

Pure file reorganization, no behavior change.
